use correct reference when using strip_component
The current implmnetation sets the environment name to the "striped"
version of the name, this is then later used as the git reference when
syncing environments.

With the following configuration
```yaml
sources:
  b4ldr:
    strip_component: "wmcs/"
    ignore_branch_prefixes:
    - production
    - ops
    - sandbox
    - pontoon
    remote: https://github.com/b4ldr/puppet-1
```

where upstream has a branch with wmcs/production then we see the
following error

  Unable to sync repo to unresolvable ref 'production' at
    /etc/puppet/code/environments/production/.git

If upstream also has a production branch then the above configuration
ends up syncing  /etc/puppet/code/environments/production/ with the
remote production branch.  even thought it should be striped via
`ignore_branch_prefixes`

This PR updates the name object to keep hold of the original name and
store it as the ref