(RK-378) Add environment_name to Puppetfile DSL
Prior to this commit, there was not an easy way to access the
environment name from the Puppetfile. This commit adds
'environment_name' to the Puppetfile DSL to allow users to access the
environment name if it is provided. The common way to access that is
'@librarian.environment_name' which will be a string containing the name
of the environment or `nil` if there was no environment provided.