(CODEMGMT-1300) Update rugged cache remote if it has changed
Prior to this commit, if the rugged cache repository had a different
remote url than the current configuration, it would not use the current
configured url. This would cause r10k deployment failures when there was
an malformed or incorrect cached remote URL. This commit changes that
behavior by updating the remote URL if the current configuration and the
cache differ. The current configuration will be used instead of the
cached URL.