(CODEMGMT-1295) Update default module git branch to `main`
The git ecosystem in general and the Puppet module ecosystem in
particular are moving from `master` as the default branch to `main`.
This commit updates r10k's default to match this shift.

Now, if a git module is defined in a Puppetfile with no ref specified,
r10k will attempt to deploy its `main` branch. If this is undesirable
for a user's module, they should specify a ref explicitly to use
instead.