(RK-388) Add splat operators to avoid warnings on Ruby 2.7
Rugged's Repository's `fetch` method expects the last set of args to be
a series of keyword arugments. Previously, we were passing these as a
hash, but this became deprecated in Ruby 2.7, and now issues a warning.
This commit updates the calls to use a splat operator instead, which
avoids the warning.