(RK-395) Allow Puppetfile override for exclude_spec
Previously, the r10k setting `exclude_spec` read in the settings
from the puppetfile first, and then overrode it with the setting
that comes from the defaults, r10k.yaml, or the cli. In order to
support PE customers who want to override the r10k.yaml setting
per module, the puppetfile setting should win out.