(RK-390) Remove hardcoded default for git ref
Previously r10k fell back to the 'main' ref if no other ref was
specified. With this patch, r10k will no longer fall back to that
ref, and error when no ref is found.

To mitigate breakage relying on the hardcoded ref, a user can
set `default_ref` in the r10k config to a ref of their
choice, mimicking the old behavior.