(RK-397) Check for module on disk before skipping implementation
Prior to this change, the puppetfile module loader would check
to see if the static version changed after an environment sync.
However, if the static version failed to deploy, this check
would prevent a deploy from re-attempting to deploy that
static version. This results in deploys returning without
error because it never re-attempts to deploy the invalid static
version.

This change just ensures that the install path for the module
also exists before skipping, so we can be reasonably certain that
what is installed on disk is the version r10k expects.