Add a wrapper file ignore list to R10K::Tarball
Some sources will wrap all content in a wrapper directory (expected),
but include a header file in the tarball root (bad). Namely, Github will
include a file named pax_global_header in the provided tarball.

This is still a valid wrapper dir, but we have to ignore the header
file.