(RK-399) Do not warn on spec changes with exclude_spec
Prior to this change, r10k would warn about local changes on the files
in the spec dir when exclude_spec was set. Since r10k was removing the
files, the warning would always be present for each git module when
exclude_spec was set. This change removes the spec files from the insync
checks so that it will not be marked as dirty if there are only changes
to the spec files. After this change, the warning about local
modifications on the spec files should not be present.