Add TOC to configuration docs
The configuration doc is structured, but it's too large to easily
understand that structure.  It can be difficult for users to find their
way around without losing the context of the documentation they
eventually find, especially for beginners.

This patch adds a hyperlinked table of contents to the configuration
doc, to clarify the document's structure and provide an index of each
option with context.

Fixes #1298