Fix possible absoulte path bug
The R10K::Tarball#unpack method previously assumed that its input was an
absolute path, but did not validate that. If a caller gave a relative
path, this could have caused problems in the algorithm the method used
to remove the wrapper dir. This commit fixes that issue by calling
File#expand_path on the method's input.

Additionally, this commit adds a comment explaining why (and how) we're
removing the wrapper dir the way we are.