(PE-35197) Cannot recover from typos in Puppetfile
Previously, if loading module metadata raised an uncaught exception
a user with automation that always used `--incremental` could not recover
from a failed deployment.

One specific case of this is a module without a proper implementation.
In that case a RuntimeError is raised. This adds RuntimeError to the
metadata loading rescue block.