(maint) Catch ScriptError, StandardError families when parsing metadata
Parsing metedata, for users who run `--incremental` with every deploy,
must not raise exceptions that end deployment. To prevent that we have
collected a large list of fairly basic execptions families.

This consolidates us to the ScriptError and StandardError exception
families, which contained all of the previous exceptions, but still
doesn't catch things like SystemExit exceptions.