(maint) Adjust jwt spec to be more relaxed
This commit updates the r10k gemspec to allow any jwt rubygem version in the 2 series after 2.2.3.
The minimum version is 2.2.3, but the only requirement on the upper bound is < 3. Note that the 3
series does not exist in jwt, this upper bounds just protects against an open ended major version.