(maint) Support different major versions of minitar
This changes us from requiring at least minitar 1.0 and instead goes to
the previous minimum verion (0.9), but instead of using a "~>" which
forbids going to a new major release, ie is equivalent to [">= 0.9", "< 1.0"]
we simply require above the previous version, not caring if we take up a
new major release. This allows us to be installed in the same version
Ruby bundle as Puppet (which currently does not support minitar >= 1.0).

Minitar has supported the "new" `::Minitar` constant since before 0.9,
so there are no code changes required.